requirements.txt
assertpy==1.0
pylint==2.5.0
pyperclip==1.8.0
pytest-cov==2.8.1
pytest==5.4.1
requests==2.23.0
vcrpy==4.0.2
